COMMERCE-7143 batch-planner-web - Reorganize names
