+++ modules/apps/batch-planner/batch-planner-web/src/main/resources/META-INF/resources/js/BatchPlannerExport.js
