+++ modules/apps/batch-planner/batch-planner-web/test/unit/api.js
@@ -17,7 +17,7 @@ import fetchMock from 'fetch-mock';
 import {
 	getExportTaskStatusURL,
 	getPollingExportStatusProcess,
-} from '../../src/main/resources/META-INF/resources/js/api';
+} from '../../src/main/resources/META-INF/resources/js/BatchPlannerExport';
 import {
 	EXPORT_PROCESS_COMPLETED,
 	EXPORT_PROCESS_FAILED,
